requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
fake-useragent==0.1.11
feedparser==6.0.10
idna==3.2
lxml==4.8.0
pydantic==1.10.11
python-dateutil==2.8.2
pytz==2023.3
requests==2.26.0
rss-parser==1.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
typing_extensions==4.7.1
urllib3==1.26.6
xmltodict==0.13.0
